Major Daemon Changes
This commit is the total effort of my work over the past 5 days. It introduces many important changes such as

1. New endpoints to list labs and get daemon info such as CPU and RAM usage along with how many sessions are running. This data will be used in the daemon selection algorithms when using multiple daemons

2. Production ready gunicorn webserver is now the default. Please make sure you have it installed on your Linux host. You can change the listen address and port in the config.json file. To start the daemon run `python3 gunicorn.py`

3. "Authentication" using a long string. Configure the connection_key in the config.json and then configure it on the frontend.

4.  POST requests instead of GET requests for daemon operations for better security and cleaner functionality

5. Updated CLI program with it's own config file. Be sure to set the daemon's URL and security key to what is configured in the daemon's config!  It is also now in it's own directory which can be copied to any comparable workstation. 


IMPORTANT: Many of these changes are for the Web UI which is a work in progress. The source code for the front-end web UI will be made available when it is ready.  As for now, you can experiment with these features using the cyberlab_cli tool.