One way Iptables reference
This is a reference you can use to allow the CyberLab host to ping/ssh (or interface with any other port) on the Lab VMs and Deny the VMs from pinging/ssh the CyberLab host. This is necessary for security especially when providing VMs with hacking tools on them. 

When this is fully implemented into Cyberlab it will be automated. For now I must work with the permissions problem for iptables. (obviously it is not safe to run CyberLab as root)