Implmented Graders
Graders are used to apply value to checkers. It returns the total points for the assignment, how many of those points were awarded for correct checkers, how many checkers were not answered (run) and how many were correct or incorrect.

Another feature that was added to the cli program was the "shutsession" option. This option does the "opposite" of "demosession". It takes the session_id as a parameter and when run it runs the checkers, runs the graders, and destroys the session. This simulates "submitting" a lab.