Updated lab JSON options
"windisk" and "lindisk" options are now just "disk". Which type of driver to use is now determined by the "OS" option which can either be "Windows" or "Linux"

You can have the DHCPv4 server assign a random IPv4 address by setting the machine's IP assignment as "x.x.x.x" instead of "x.x.x.n" where n is a number in the ipassignments for a network.